+++ src/equations/average.c
@@ -13,22 +13,22 @@ void *equations_average(const Equations *eqns, const char *entity, const void *v
     scalar *volume = eqns->mesh->cells.volume;
     scalar sum_volume = eqns->mesh->cells.sum_volume;
 
-    int num = eqns->mesh->entities.num;
+    long num = eqns->mesh->entities.num;
     Name *name = eqns->mesh->entities.name;
-    int *cell_off = eqns->mesh->entities.cell_off;
+    long *cell_off = eqns->mesh->entities.cell_off;
 
-    int stride = eqns->variables.stride;
+    long stride = eqns->variables.stride;
     const scalar(*variable)[stride] = variable_;
 
-    for (int i = 0; i < num; i++) {
+    for (long i = 0; i < num; i++) {
         if (!strcmp(name[i], entity)) {
             scalar *average = arena_calloc(stride, sizeof(*average));
-            for (int j = cell_off[i]; j < cell_off[i + 1]; j++) {
-                for (int k = 0; k < stride; k++) {
+            for (long j = cell_off[i]; j < cell_off[i + 1]; j++) {
+                for (long k = 0; k < stride; k++) {
                     average[k] += volume[j] * variable[j][k];
                 }
             }
-            for (int j = 0; j < stride; j++) {
+            for (long j = 0; j < stride; j++) {
                 average[j] = sync_fsum(average[j]) / sum_volume;
             }
             return average;
