+++ src/equations/gradient.c
@@ -8,24 +8,24 @@ void *equations_gradient(const Equations *eqns, void *variable_)
 {
     assert(eqns && variable_);
 
-    int num = eqns->mesh->cells.num;
+    long num = eqns->mesh->cells.num;
 
-    int num_faces = eqns->mesh->faces.num;
-    int num_inner = eqns->mesh->faces.num_inner;
-    int off_ghost = eqns->mesh->faces.off_ghost;
+    long num_faces = eqns->mesh->faces.num;
+    long num_inner = eqns->mesh->faces.num_inner;
+    long off_ghost = eqns->mesh->faces.off_ghost;
     Adjacent *cell = eqns->mesh->faces.cell;
     vector *weight = eqns->mesh->faces.weight;
 
-    int stride = eqns->variables.stride;
+    long stride = eqns->variables.stride;
     scalar(*variable)[stride] = variable_;
     vector(*gradient)[stride] = arena_calloc(num, sizeof(*gradient));
 
     Request req = sync_variables(eqns, variable, stride);
 
-    for (int i = 0; i < num_inner; i++) {
-        int left = cell[i].left;
-        int right = cell[i].right;
-        for (int j = 0; j < stride; j++) {
+    for (long i = 0; i < num_inner; i++) {
+        long left = cell[i].left;
+        long right = cell[i].right;
+        for (long j = 0; j < stride; j++) {
             scalar diff = variable[right][j] - variable[left][j];
             gradient[left][j].x += diff * weight[i].x;
             gradient[left][j].y += diff * weight[i].y;
@@ -35,10 +35,10 @@ void *equations_gradient(const Equations *eqns, void *variable_)
             gradient[right][j].z += diff * weight[i].z;
         }
     }
-    for (int i = num_inner; i < off_ghost; i++) {
-        int left = cell[i].left;
-        int right = cell[i].right;
-        for (int j = 0; j < stride; j++) {
+    for (long i = num_inner; i < off_ghost; i++) {
+        long left = cell[i].left;
+        long right = cell[i].right;
+        for (long j = 0; j < stride; j++) {
             scalar diff = variable[right][j] - variable[left][j];
             gradient[left][j].x += diff * weight[i].x;
             gradient[left][j].y += diff * weight[i].y;
@@ -48,10 +48,10 @@ void *equations_gradient(const Equations *eqns, void *variable_)
 
     sync_wait(eqns, req.recv);
 
-    for (int i = off_ghost; i < num_faces; i++) {
-        int left = cell[i].left;
-        int right = cell[i].right;
-        for (int j = 0; j < stride; j++) {
+    for (long i = off_ghost; i < num_faces; i++) {
+        long left = cell[i].left;
+        long right = cell[i].right;
+        for (long j = 0; j < stride; j++) {
             scalar diff = variable[right][j] - variable[left][j];
             gradient[left][j].x += diff * weight[i].x;
             gradient[left][j].y += diff * weight[i].y;
