+++ src/equations/norm.c
@@ -12,12 +12,12 @@ void *equations_norm(const Equations *eqns, scalar time)
 
     Arena save = arena_save();
 
-    int num = eqns->mesh->cells.num_inner;
+    long num = eqns->mesh->cells.num_inner;
     vector *center = eqns->mesh->cells.center;
     scalar *volume = eqns->mesh->cells.volume;
     scalar sum_volume = eqns->mesh->cells.sum_volume;
 
-    int stride = eqns->variables.stride;
+    long stride = eqns->variables.stride;
     scalar(*variable)[stride] = eqns->variables.data;
     scalar *property = eqns->properties.data;
     Compute *compute = eqns->user.compute;
@@ -26,14 +26,14 @@ void *equations_norm(const Equations *eqns, scalar time)
     scalar *norm = arena_calloc(stride, sizeof(*norm));
     scalar *user = arena_calloc(stride, sizeof(*user));
 
-    for (int i = 0; i < num; i++) {
+    for (long i = 0; i < num; i++) {
         compute(user, property, center[i], time, variable[i]);
         conserved(user, property);
-        for (int j = 0; j < stride; j++) {
+        for (long j = 0; j < stride; j++) {
             norm[j] += volume[i] * sq(user[j] - variable[i][j]);
         }
     }
-    for (int i = 0; i < stride; i++) {
+    for (long i = 0; i < stride; i++) {
         norm[i] = sqrt(sync_fsum(norm[i]) / sum_volume);
     }
