+++ src/equations/restart.c
@@ -3,7 +3,7 @@
 #include "equations.h"
 #include "teal/option.h"
 
-void equations_restart(const Equations *eqns, scalar *time, int *index)
+void equations_restart(const Equations *eqns, scalar *time, long *index)
 {
     assert(eqns && time && index);
     if (!option.restart) {
